Add files via upload
TODO LIST APP


Overview


The TODO LIST APP is a simple web application that allows users to create, read, update, and delete (CRUD) todo items. The app is built using HTML, CSS, and JavaScript, and is designed to be a basic example of a todo list app.

features


* Add new todo items by typing in the input field and clicking the "Add Todo" button.
* Delete todo items by clicking the "X" button next to each todo item.
* The todo list is rendered dynamically based on the `todos` array.
* The todo list is updated in real-time as todos are added or deleted.

Requirements

* HTML5
* CSS3
* JavaScript (ES6+)
* A modern web browser (e.g. Google Chrome, Mozilla Firefox)

Installation
------------

1. Clone the repository using Git: `git clone https://github.com/your-username/todo-list-app.git`
2. Open the `index.html` file in a web browser to run the app.

Use

1. Type in the input field to create a new todo item.
2. Click the "Add Todo" button to add the todo item to the list.
3. Click the "X" button next to each todo item to delete it.
4. The todo list will be updated in real-time as todos are added or deleted.

**Code Structure**


* `index.html`: The main HTML file that contains the app's structure and content.
* `style.css`: The CSS file that styles the app's layout and design.
* `script.js`: The JavaScript file that contains the app's logic and functionality.

**License**

The TODO LIST APP is licensed under the MIT License.

**Contributing**
------------

Contributions are welcome! If you'd like to contribute to the TODO LIST APP, please fork the repository and submit a pull request.

**Issues**

If you encounter any issues with the TODO LIST APP, please submit an issue on the GitHub repository.

**Author**


The TODO LIST APP was created by [Your Name].

**Acknowledgments**


The TODO LIST APP was built using the following resources:

* HTML5
* CSS3
* JavaScript (ES6+)
* GitHub